mem_barrier

Enum BarrierKind

#[non_exhaustive]
pub enum BarrierKind {
    Mmio,
    Smp,
    Dma,
    Compiler,
}

The kind of a memory barrier.

This enum determines the strength or flavor of the memory barrier.

Current implementation

On x86, this does not affect instruction generation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Mmio

MMIO.

This is the strongest kind of barrier. It enforces ordering on memory accesses as well as on MMIO-based device I/O.

Corresponding functions

This kind of barrier corresponds to the mandatory mb, rmb, and wmb Linux functions.

Current implementation

On Arm, this runs a DSB instruction; see Data Synchronization Barrier.

Smp

SMP.

This kind of barrier enforces ordering on memory across an SMP system. It is also suitable to enforce ordering on memory shared from a single-CPU VM guest with an SMP host.

Corresponding functions

This kind of barrier corresponds to the VM-guest-flavoured virt_mb, virt_rmb, and virt_wmb Linux functions. Those functions are equivalent to the SMP-flavoured smp_mb, smp_rmb, and smp_wmb Linux functions when SMP support is turned on.

Note that this kind of barrier does not change its behavior based on the build configuration.

Current implementation

On Arm, this runs a DMB instruction; see Data Memory Barrier.

Dma

DMA.

This kind of barrier enforces ordering on memory accessed by the CPU and DMA-capable devices.

Corresponding functions

This kind of barrier corresponds to the DMA-flavoured dma_mb, dma_rmb, and dma_wmb Linux functions.

Current implementation

On Arm, this runs a DMB instruction; see Data Memory Barrier.

Examples

This example is inspired by the Linux dma_rmb and dma_wmb example.

use mem_barrier::{BarrierKind, BarrierType, mem_barrier};

if !desc.device_owns_memory() {
    // Don't read until we own the descriptor.
    mem_barrier(BarrierKind::Dma, BarrierType::Read);

    // Read/modify data
    read_data = desc.data();
    desc.set_data(write_data);

    // Flush modifications.
    mem_barrier(BarrierKind::Dma, BarrierType::Write);

    // Give the descriptor ownership back to the device.
    desc.set_device_owns_memory(true);

    // Notify the device.
    device.notify();
}

Compiler

Compiler.

This kind of barrier does not run any CPU instructions. Instead, it only prevents the compiler from moving memory accesses through the barrier.

Corresponding functions

This kind of barrier corresponds to the barrier Linux function.